Move ListView into its own file
